order form create task

+++ app/Http/Controllers/UserContactController.php
@@ -6,9 +6,19 @@ use App\User;
 use App\UserContact;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
+
 
 class UserContactController extends Controller
 {
+    public $user;
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::user();
+            return $next($request);
+        });
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -16,6 +26,10 @@ class UserContactController extends Controller
      */
     public function create(UserContact $userContact,$slug)
     {
+        if (is_null($this->user) ||  !$this->user->can('user.contact.create')) {
+            $message = 'You are not allowed to access this page !';
+            return view('errors.403', compact('message'));
+        }
         $data['lists']=User::where('slug',$slug)->first();
         return view ('admin.shipper.contact.create',$data)->with('data',$userContact);
     }
@@ -28,6 +42,10 @@ class UserContactController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) ||  !$this->user->can('user.contact.store')) {
+            $message = 'You are not allowed to access this page !';
+            return view('errors.403', compact('message'));
+        }
         $request->validate([
             'contact' => 'required|max:255',
             'email' => 'required|unique:user_contacts',
@@ -66,6 +84,10 @@ class UserContactController extends Controller
      */
     public function edit(UserContact $userContact,$slug)
     {
+        if (is_null($this->user) ||  !$this->user->can('user.contact.edit')) {
+            $message = 'You are not allowed to access this page !';
+            return view('errors.403', compact('message'));
+        }
         $data['lists']=UserContact::where('slug',$slug)->first();
         return view ('admin.shipper.contact.edit',$data);
 
@@ -80,6 +102,10 @@ class UserContactController extends Controller
      */
     public function update(Request $request)
     {
+        if (is_null($this->user) ||  !$this->user->can('user.contact.update')) {
+            $message = 'You are not allowed to access this page !';
+            return view('errors.403', compact('message'));
+        }
         $data = UserContact::where("id",$request->id)->first();
         $data->update([
             $data->contact=$request->contact,
@@ -100,6 +126,10 @@ class UserContactController extends Controller
      */
     public function destroy($slug)
     {
+        if (is_null($this->user) ||  !$this->user->can('user.contact.destroy')) {
+            $message = 'You are not allowed to access this page !';
+            return view('errors.403', compact('message'));
+        }
         $data= UserContact::where("slug",$slug)->first();
         $data->delete();
         return redirect()->back();
